May 23: Really should commit more
So far, I mainly added a few more fractals, and some visual flare. For
starters, I added an oscillation function for the boxes class, which
simply makes it move around. After that, I added a few more fractals.
One is the Menger Sponge, which I got the idea for from a friend,
although the actual code was done by me. I also added a tree fractal
and another custom fractal. I also tried to change up the visuals, and
added a field of view variable, that changes slightly when moving
forward, which is a small visual detail.

+Added Menger Sponge fractal
+Added tree fractal
+Added custom fractal
+Added movement zoom
+Added option for oscillating platforms
*changed colour scheme