8 red coins
So far, the most major addition is the that of a score system, based on
collecting various red coins strewn around the map. While I tried to
mess around with integrating it into a HUD, it required me to install
another camera library, which went about as well as a whale trying to
ride a unicycle. I again changed the colour scheme, this time to
something a bit more in the vain of “Sin City”. Another major change
the inclusion of a complete jump that functions like actual physics
apply to it. Finally, I added a ceiling collision to my box class.

+Fully fleshed out jumping mechanics.
+Complete collisions
+Score system
+New colours scheme

-Removed old colour scheme
-Removed Herobrine